ocean-shortterm/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile requirements.in
#
aiofiles==23.2.1
    # via gradio
aiohttp==3.9.1
    # via langchain
aiosignal==1.3.1
    # via aiohttp
altair==5.2.0
    # via gradio
annotated-types==0.6.0
    # via pydantic
anyio==3.7.1
    # via
    #   fastapi
    #   httpx
    #   langchain-core
    #   openai
    #   starlette
attrs==23.1.0
    # via
    #   aiohttp
    #   jsonschema
    #   referencing
certifi==2023.11.17
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via
    #   typer
    #   uvicorn
colorama==0.4.6
    # via typer
contourpy==1.2.0
    # via matplotlib
cycler==0.12.1
    # via matplotlib
dataclasses-json==0.6.3
    # via langchain
distro==1.8.0
    # via openai
fastapi==0.104.1
    # via gradio
ffmpy==0.3.1
    # via gradio
filelock==3.13.1
    # via huggingface-hub
fonttools==4.46.0
    # via matplotlib
frozenlist==1.4.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2023.12.1
    # via
    #   gradio-client
    #   huggingface-hub
gradio==4.8.0
    # via -r requirements.in
gradio-client==0.7.1
    # via gradio
greenlet==3.0.2
    # via sqlalchemy
h11==0.14.0
    # via
    #   httpcore
    #   uvicorn
httpcore==1.0.2
    # via httpx
httpx==0.25.2
    # via
    #   gradio
    #   gradio-client
    #   openai
huggingface-hub==0.19.4
    # via
    #   gradio
    #   gradio-client
idna==3.6
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
importlib-resources==6.1.1
    # via gradio
jinja2==3.1.2
    # via
    #   altair
    #   gradio
json-repair==0.4.5
    # via -r requirements.in
jsonpatch==1.33
    # via
    #   langchain
    #   langchain-core
jsonpointer==2.4
    # via jsonpatch
jsonschema==4.20.0
    # via altair
jsonschema-specifications==2023.11.2
    # via jsonschema
kiwisolver==1.4.5
    # via matplotlib
langchain==0.0.348
    # via -r requirements.in
langchain-core==0.0.12
    # via langchain
langsmith==0.0.69
    # via
    #   langchain
    #   langchain-core
markdown-it-py==3.0.0
    # via rich
markupsafe==2.1.3
    # via
    #   gradio
    #   jinja2
marshmallow==3.20.1
    # via dataclasses-json
matplotlib==3.8.2
    # via gradio
mdurl==0.1.2
    # via markdown-it-py
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.0.0
    # via typing-inspect
numpy==1.26.2
    # via
    #   altair
    #   contourpy
    #   gradio
    #   langchain
    #   matplotlib
    #   pandas
openai==1.3.8
    # via -r requirements.in
orjson==3.9.10
    # via gradio
packaging==23.2
    # via
    #   altair
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   langchain-core
    #   marshmallow
    #   matplotlib
pandas==2.1.4
    # via
    #   altair
    #   gradio
pillow==10.1.0
    # via
    #   gradio
    #   matplotlib
pydantic==2.5.2
    # via
    #   fastapi
    #   gradio
    #   langchain
    #   langchain-core
    #   langsmith
    #   openai
pydantic-core==2.14.5
    # via pydantic
pydub==0.25.1
    # via gradio
pygments==2.17.2
    # via rich
pyparsing==3.1.1
    # via matplotlib
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
python-dotenv==1.0.0
    # via -r requirements.in
python-multipart==0.0.6
    # via gradio
pytz==2023.3.post1
    # via pandas
pyyaml==6.0.1
    # via
    #   gradio
    #   huggingface-hub
    #   langchain
    #   langchain-core
referencing==0.32.0
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2023.10.3
    # via tiktoken
requests==2.31.0
    # via
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   langchain
    #   langchain-core
    #   langsmith
    #   tiktoken
rich==13.7.0
    # via typer
rpds-py==0.13.2
    # via
    #   jsonschema
    #   referencing
semantic-version==2.10.0
    # via gradio
shellingham==1.5.4
    # via typer
six==1.16.0
    # via python-dateutil
sniffio==1.3.0
    # via
    #   anyio
    #   httpx
    #   openai
sortedcontainers==2.4.0
    # via -r requirements.in
sqlalchemy==2.0.23
    # via langchain
starlette==0.27.0
    # via fastapi
tenacity==8.2.3
    # via
    #   langchain
    #   langchain-core
tiktoken==0.5.2
    # via -r requirements.in
tomlkit==0.12.0
    # via gradio
toolz==0.12.0
    # via altair
tqdm==4.66.1
    # via
    #   huggingface-hub
    #   openai
typer[all]==0.9.0
    # via
    #   gradio
    #   typer
typing-extensions==4.9.0
    # via
    #   fastapi
    #   gradio
    #   gradio-client
    #   huggingface-hub
    #   openai
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typer
    #   typing-inspect
typing-inspect==0.9.0
    # via dataclasses-json
tzdata==2023.3
    # via pandas
urllib3==2.1.0
    # via requests
uvicorn==0.24.0.post1
    # via gradio
websockets==11.0.3
    # via gradio-client
yarl==1.9.4
    # via aiohttp
